Optionally avoid wordexp when not needed
The musl implementation of wordexp launches a shell to do a printf for each
command line parameter we try to expand, which can be pretty noisy in telemetry.
This shuts it off globally via a config flag.

It might be nice to be able to enable/disable on a finer-grained basis, but
doing the config for that seemed like more trouble than it's worth right now.